Bug 1283056 - Cast the enum to int to fix a bustage on a CLOSED TREE.
MozReview-Commit-ID: 8g6gJpz66cu

+++ dom/media/webaudio/AudioContext.cpp
@@ -854,7 +854,7 @@ AudioContext::OnStateChanged(void* aPromise, AudioContextState aNewState)
       (mAudioContextState == aNewState))) {
     fprintf(stderr,
             "Invalid transition: mAudioContextState: %d -> aNewState %d\n",
-            mAudioContextState, aNewState);
+            static_cast<int>(mAudioContextState), static_cast<int>(aNewState));
     MOZ_ASSERT(false);
   }
